ada: Add Incomplete_Type_OK formal to Process_Subtype
Before this patch, Process_Subtype looked at the parent of its argument
to determine whether it was called in a context where it was OK for the
subtype mark to refer to the incomplete view of a type. This patch adds
a new formal so that it becomes the responsibility of the caller in
order to make the code simpler overall.

The only change in the behavior of the compiler is that the messages
emitted in some error situations with the -gnatf switch are slightly
different.

gcc/ada/ChangeLog:

	* sem_ch3.ads (Process_Subtype): New formal.
	* sem_ch3.adb (Process_Subtype): Likewise.
	(Analyze_Subtype_Declaration, Access_Type_Declaration): Use new
	formal.