move global data to symbol_table_init
The following avoids early runtime initialization of cbl_field_t
objects which, when using tree to represent the current _Float128
data, segfaults as tree data like float128_type_node is not yet
initialized.  The solution is to move the global data to
symbol_table_init which is the only user and it already has one
such instance locally, the 'constants' array.

	* symbols.cc (empty_float, empty_comp5, empty_literal,
	empty_conditional, debug_registers, special_registers): Move
	global cbl_field_t typed data to ...
	(symbol_table_init): ... local scope here.