ada: Error on subtype with static predicate used in case_expression
The compiler improperly flags an error on the use of a subtype with a
static predicate as a choice in a case expression alternative, complaining
that the subtype has a nonstatic predicate. The fix for this is to add
a test for the subtype not having a static predicate.

gcc/ada/ChangeLog:

	* einfo.ads: Revise comment about Dynamic_Predicate flag to make it
	more accurate.
	* sem_case.adb (Check_Choices): Test "not Has_Static_Predicate_Aspect"
	as additional guard for error about use of subtype with nonstatic
	predicate as a case choice. Improve related error message.