AVR: libgcc: Properly exclude object files for AVRrc.
There are many objects / functions that are not available on AVRrc,
the reduced core.  The old way to exclude some objects for AVRrc
did not work properly since it tested for MULTIFLAGS.
This does not work for, say MULTIFLAGS = "-mmcu=avrtiny -mdouble=64".
This patch uses $(findstring avrtiny,$(MULTIDIR)) in the condition.

libgcc/
	* config/avr/t-avr (LIB1ASMFUNCS, LIB2FUNCS_EXCLUDE):
	Properly handle avrtiny.
libgcc/config/avr/libf7/
	* t-libf7 (libgcc-objects): Only add objects when building
	for non-AVRrc.